Rename: Product 패키지 병합
