+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -7,11 +7,11 @@ import java.util.Arrays;
 import java.util.Collections;
 import kitchenpos.menu.dao.MenuDao;
 import kitchenpos.menu.dao.MenuGroupDao;
-import kitchenpos.product.dao.ProductDao;
+import kitchenpos.menu.dao.ProductDao;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import org.junit.jupiter.api.BeforeEach;
